Make Client depends on web3 and manage NodeProcess explicitly
