+++ tests/golem/transactions/ethereum/test_ethereumtransactionsystem.py
@@ -61,14 +61,13 @@ class TestEthereumTransactionSystem(TestWithDatabase, LogTestCase,
             mock_is_service_running.return_value = False
             e = EthereumTransactionSystem(self.tempdir, PRIV_KEY)
             assert e.payment_processor._loopingCall.start.called
-            assert e._sci._geth_client.node.start.called
+            assert e._node.start.called
 
             mock_is_service_running.return_value = False
             e.stop()
-            assert e._sci._geth_client.node is None
+            assert e._node.stop.called
             assert not e.payment_processor._loopingCall.stop.called
 
             mock_is_service_running.return_value = True
             e.stop()
-            assert e._sci._geth_client.node is None
             assert e.payment_processor._loopingCall.stop.called
